requirements.txt
acrylic
streamlit
pandas
google-api-python-client
st-gsheets-connection
